move api wrapper to new ts file
